+++ next.config.ts
@@ -10,6 +10,7 @@ const nextConfig: NextConfig = {
             { protocol: 'https', hostname: 'is2.4chan.org' },
             { protocol: 'https', hostname: 'easychan.net' },
             { protocol: 'https', hostname: 'mokachan.cafe' },
+            { protocol: 'https', hostname: '2ch.org' },
         ],
     },
     allowedDevOrigins: process.env.ALLOWED_DEV_ORIGINS